# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A spaced repetition API for tracking Leetcode practice sessions.",
        "title": "Leetcode Buddy API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
        "/dashboard": {
            "get": {
                "description": "Returns problems due for review and problems with the lowest FSRS stability",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Dashboard summary",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Number of items per table (1–50, default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DashboardData"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/problems": {
            "get": {
                "description": "Returns a paginated list of Leetcode problems, optionally filtered by topics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problems"
                ],
                "summary": "List problems",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by topic names",
                        "name": "topics",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProblemPage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/problems/submissions": {
            "get": {
                "description": "Returns submissions across all problems, optionally filtered by problem IDs",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "List all submissions",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "collectionFormat": "csv",
                        "description": "Filter by problem IDs",
                        "name": "problemId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProblemSubmissionWithDetailsPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Saves a new attempt for a Leetcode problem",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Record a new submission",
                "parameters": [
                    {
                        "description": "Submission details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.saveProblemSubmissionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/problems/submissions/import": {
            "post": {
                "description": "Accepts a multipart/form-data .xlsx file and bulk-imports problem submissions. Each row must have columns: \"Problem #\", \"Date\" (YYYY-MM-DD), \"Confidence\" (1-4), \"Language\" (python only).",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "submissions"
                ],
                "summary": "Import submissions from Excel",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Excel file (.xlsx)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ImportSubmissionsResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/problems/topics": {
            "get": {
                "description": "Returns every topic that exists across all problems",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problems"
                ],
                "summary": "List all problem topics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/problems/{id}": {
            "get": {
                "description": "Returns a single problem including its topics",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problems"
                ],
                "summary": "Get a problem by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Problem ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProblemDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/problems/{id}/submissions": {
            "get": {
                "description": "Returns all submissions for the problem identified by id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "problems"
                ],
                "summary": "Get submissions for a problem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Problem ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProblemSubmissionPage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controllers.saveProblemSubmissionRequest": {
            "type": "object",
            "required": [
                "confidenceLevel",
                "problemId"
            ],
            "properties": {
                "confidenceLevel": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "problemId": {
                    "type": "integer",
                    "minimum": 1
                },
                "timeTaken": {
                    "type": "integer"
                }
            }
        },
...
        "models.DashboardData": {
            "type": "object",
            "properties": {
                "due": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProblemReviewItem"
                    }
                },
                "lowStability": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProblemReviewItem"
                    }
                },
                "overdueCount": {
                    "type": "integer"
                }
            }
        },
        "models.ImportSubmissionRowError": {
            "type": "object",
            "properties": {
                "reason": {
                    "type": "string"
                },
                "row": {
                    "type": "integer"
                }
            }
        },
        "models.ImportSubmissionsResult": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ImportSubmissionRowError"
                    }
                },
                "imported": {
                    "type": "integer"
                },
                "skipped": {
                    "type": "integer"
                }
            }
        },
        "models.Problem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/models.ProblemDifficulty"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.ProblemDetail": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/models.ProblemDifficulty"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "stability": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "topics": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.ProblemDifficulty": {
            "type": "string",
            "enum": [
                "easy",
                "medium",
                "hard"
            ],
            "x-enum-varnames": [
                "EasyDifficulty",
                "MediumDifficulty",
                "HardDifficulty"
            ]
        },
        "models.ProblemPage": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Problem"
                    }
                }
            }
        },
        "models.ProblemReviewItem": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "$ref": "#/definitions/models.ProblemDifficulty"
                },
                "due": {
                    "type": "string"
                },
                "problemId": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "stability": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ProblemSubmission": {
            "type": "object",
            "properties": {
                "confidenceLevel": {
                    "$ref": "#/definitions/models.ConfidenceLevel"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "$ref": "#/definitions/models.SubmissionLanguage"
                },
                "problemId": {
                    "type": "integer"
                },
                "timeTaken": {
                    "type": "integer"
                }
            }
        },
        "models.ProblemSubmissionPage": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProblemSubmission"
                    }
                }
            }
        },
        "models.ProblemSubmissionWithDetails": {
            "type": "object",
            "properties": {
                "confidenceLevel": {
                    "$ref": "#/definitions/models.ConfidenceLevel"
                },
                "id": {
                    "type": "string"
                },
                "problem": {
                    "type": "object",
                    "properties": {
                        "description": {
                            "type": "string"
                        },
                        "difficulty": {
                            "$ref": "#/definitions/models.ProblemDifficulty"
                        },
                        "id": {
                            "type": "integer"
                        },
                        "title": {
                            "type": "string"
                        }
                    }
                },
                "submittedAt": {
                    "type": "string"
                },
                "timeTaken": {
                    "type": "integer"
                }
            }
        },
        "models.ProblemSubmissionWithDetailsPage": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProblemSubmissionWithDetails"
                    }
                }
            }
        },
...
    }
}